+++ src/types/PMatrix.java
@@ -47,10 +47,10 @@ public interface PMatrix {
   public PMatrix get();  
 
   /**
-   * Copies the matrix contents into a float array.
+   * Copies the matrix contents into a double array.
    * If target is null (or not the correct size), a new array will be created.
    */
-  public float[] get(float[] target);
+  public double[] get(double[] target);
   
   
   /**
@@ -62,46 +62,46 @@ public interface PMatrix {
    * Set the contents of this matrix to the contents of source. Fills the
    * matrix left-to-right, starting in the top row.
    */
-  public void set(float[] source);
+  public void set(double[] source);
 
   /**
    * Set the matrix content to this 2D matrix or its 3D equivalent.
    */
-  public void set(float m00, float m01, float m02, 
-                  float m10, float m11, float m12);
+  public void set(double m00, double m01, double m02, 
+                  double m10, double m11, double m12);
 
   /**
    * Set the matrix content to the 3D matrix supplied, if this matrix is 3D.
    */
-  public void set(float m00, float m01, float m02, float m03,
-                  float m10, float m11, float m12, float m13,
-                  float m20, float m21, float m22, float m23,
-                  float m30, float m31, float m32, float m33);
+  public void set(double m00, double m01, double m02, double m03,
+                  double m10, double m11, double m12, double m13,
+                  double m20, double m21, double m22, double m23,
+                  double m30, double m31, double m32, double m33);
 
   
-  public void translate(float tx, float ty);
+  public void translate(double tx, double ty);
   
-  public void translate(float tx, float ty, float tz);
+  public void translate(double tx, double ty, double tz);
 
-  public void rotate(float angle);
+  public void rotate(double angle);
 
-  public void rotateX(float angle);
+  public void rotateX(double angle);
 
-  public void rotateY(float angle);
+  public void rotateY(double angle);
 
-  public void rotateZ(float angle);
+  public void rotateZ(double angle);
 
-  public void rotate(float angle, float v0, float v1, float v2);
+  public void rotate(double angle, double v0, double v1, double v2);
 
-  public void scale(float s);
+  public void scale(double s);
 
-  public void scale(float sx, float sy);
+  public void scale(double sx, double sy);
 
-  public void scale(float x, float y, float z);
+  public void scale(double x, double y, double z);
   
-  public void shearX(float angle);
+  public void shearX(double angle);
   
-  public void shearY(float angle);
+  public void shearY(double angle);
 
   /**
    * Multiply this matrix by another.
@@ -121,16 +121,16 @@ public interface PMatrix {
   /**
    * Multiply this matrix by another.
    */
-  public void apply(float n00, float n01, float n02, 
-                    float n10, float n11, float n12);
+  public void apply(double n00, double n01, double n02, 
+                    double n10, double n11, double n12);
 
   /**
    * Multiply this matrix by another.
    */
-  public void apply(float n00, float n01, float n02, float n03,
-                    float n10, float n11, float n12, float n13,
-                    float n20, float n21, float n22, float n23,
-                    float n30, float n31, float n32, float n33);
+  public void apply(double n00, double n01, double n02, double n03,
+                    double n10, double n11, double n12, double n13,
+                    double n20, double n21, double n22, double n23,
+                    double n30, double n31, double n32, double n33);
 
   /**
    * Apply another matrix to the left of this one.
@@ -150,16 +150,16 @@ public interface PMatrix {
   /**
    * Apply another matrix to the left of this one.
    */
-  public void preApply(float n00, float n01, float n02, 
-                       float n10, float n11, float n12);
+  public void preApply(double n00, double n01, double n02, 
+                       double n10, double n11, double n12);
 
   /**
    * Apply another matrix to the left of this one. 3D only.
    */
-  public void preApply(float n00, float n01, float n02, float n03,
-                       float n10, float n11, float n12, float n13,
-                       float n20, float n21, float n22, float n23,
-                       float n30, float n31, float n32, float n33);
+  public void preApply(double n00, double n01, double n02, double n03,
+                       double n10, double n11, double n12, double n13,
+                       double n20, double n21, double n22, double n23,
+                       double n30, double n31, double n32, double n33);
 
   
   /**
@@ -176,15 +176,15 @@ public interface PMatrix {
    * Supplying and recycling a target array improves performance, so it's
    * recommended if you call this many times in draw().
    */
-  public float[] mult(float[] source, float[] target);
+  public double[] mult(double[] source, double[] target);
   
   
-//  public float multX(float x, float y);
-//  public float multY(float x, float y);
+//  public double multX(double x, double y);
+//  public double multY(double x, double y);
   
-//  public float multX(float x, float y, float z);
-//  public float multY(float x, float y, float z);
-//  public float multZ(float x, float y, float z);  
+//  public double multX(double x, double y, double z);
+//  public double multY(double x, double y, double z);
+//  public double multZ(double x, double y, double z);  
   
   
   /**
@@ -204,5 +204,5 @@ public interface PMatrix {
   /**
    * @return the determinant of the matrix
    */
-  public float determinant();
+  public double determinant();
 }
